- refactored test_sanity and added integrity test file
